fix: update and delete

+++ owner/delete.php
@@ -1,70 +1,32 @@
-<?php 
+<?php
+// delete.php
 $property_id='';
 include("../config/config.php");
 
-if(isset($_POST['delete_property'])){
-	delete_property();
+if(isset($_POST['delete_property']) && isset($_POST['property_id'])){
+    delete_property();
 }
-
 
 function delete_property(){
-	global $db,$property_id;
-		
-$property_id=$_GET['property_id'];
+    global $db, $property_id;
+    
+    $property_id=$_POST['property_id'];
 
-$sql="DELETE from property_photo where property_id='$property_id'";
-$query=mysqli_query($db,$sql);
+    $sql="DELETE from property_photo where property_id='$property_id'";
+    $query=mysqli_query($db,$sql);
 
-if($query){
-	$sql2="DELETE from review where property_id='$property_id'";
-$query2=mysqli_query($db,$sql2);
+    if($query){
+        $sql2="DELETE from review where property_id='$property_id'";
+        $query2=mysqli_query($db,$sql2);
 
-$sql3="DELETE from add_property where property_id='$property_id'";
-$query3=mysqli_query($db,$sql3);
-if($query3){
-			
+        $sql3="DELETE from add_property where property_id='$property_id'";
+        $query3=mysqli_query($db,$sql3);
+
+        if($query3){
+            
+            header("Location: owner-index.php?success=1");
+            exit();
+        }
+    }
+}
 ?>
-
-<style>
-.alert {
-  padding: 20px;
-  background-color: #DC143C;
-  color: white;
-}
-
-.closebtn {
-  margin-left: 15px;
-  color: white;
-  font-weight: bold;
-  float: right;
-  font-size: 22px;
-  line-height: 20px;
-  cursor: pointer;
-  transition: 0.3s;
-}
-
-.closebtn:hover {
-  color: black;
-}
-</style>
-<script>
-	window.setTimeout(function() {
-    $(".alert").fadeTo(1000, 0).slideUp(500, function(){
-        $(this).remove(); 
-    });
-}, 2000);
-</script>
-<div class="container">
-<div class="alert" role='alert'>
-  <span class="closebtn" onclick="this.parentElement.style.display='none';">&times;</span> 
-  <center><strong>Your Product has been deleted.</strong></center>
-</div></div>
-
-
-<?php
-
-}
-}}
-
-
- ?>
